Refactoring Measurement Readers extracted abstract superclass to encapsulate common functionality handled by copy-paste before
